Add shasum verification. Improve the DN string parsing.
Simply looking for the correct provisioning profile based on the code signing
identity string is ambiguous and may lead to unwanted errors. We now match
using the shasum of the certificate itself which is much more deterministic.

Change the parsing of the DN string which had errors in case the code signing
identity contained "," in its name. We are now able to correctly handle those
cases. The heuristic is still using some basic regex, so there's still room
for errors.